+++ lib/models/test.dart
@@ -1,0 +1,1 @@
+class Test {}
