+++ lib/models/user.dart
@@ -1,6 +1,6 @@
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:http/http.dart' as http;
-import 'globals.dart' as globals;
+import '../globals.dart' as globals;
 import 'resolution.dart';
 import 'dart:convert';
